requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1.10
django-paypal==1.0.0
django-pyodbc-azure==2.1.0.0
idna==2.8
pyodbc==4.0.26
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3